Fixed CENTER depth call+save data
dpeth_val needs to call [ROW][Col] that is why it was saying out of range so often.

Created an empty array, to which tvec data is appended either for 1 or multiple markers. Needed a #markers=1 condition or else it wouldn't add.

+ the id value is added to it
made us of dimensional reduction

created a pandas data frame

succcessfully outputs to excel.

Still need to look into the analyses